Dynamic_Prod5/requirements.txt
# Core Python Libraries
python-dotenv>=0.21.0

# Web and Network
requests>=2.28.1
urllib3>=1.26.12

# Data Processing
pandas>=1.5.1
numpy>=1.23.4

# Document Processing
pdf2image>=1.16.0
Pillow>=9.3.0
pytesseract>=0.3.10
python-dateutil>=2.8.2

# Machine Learning and AI
openai==0.28
scikit-learn>=1.1.3

# Elasticsearch
elasticsearch==8.9.0

# Language Processing
spacy>=3.4.3

# Logging and Utilities
colorlog>=6.7.0

# Error Handling and Type Checking
typing-extensions>=4.4.0

# Image Processing
opencv-python-headless>=4.7.0.72

# PDF Processing
PyPDF2>=3.0.1

# Advanced Text Processing
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# Cryptography and Security
cryptography>=39.0.2

# Development and Debugging
ipython>=8.6.0
pdbpp>=0.10.3

# Testing
pytest>=7.2.0
pytest-cov>=4.0.0
mock>=4.0.3

# Code Quality
flake8>=5.0.4
black>=22.10.0
isort>=5.10.1

# Optional: For potential API development
fastapi==0.100.0
uvicorn==0.22.0

# Optional: For advanced data manipulation
polars>=0.16.0

# Optional: For advanced date handling
arrow>=1.2.3

# Specific library for advanced text extraction
unstructured>=0.6.8

Pillow>=9.3.0

pydantic
python-multipart
